Remove old Authorization implementation
