+++ sources/Google.Solutions.IapDesktop.Application/Profile/Auth/Authorization.cs
@@ -1,5 +1,5 @@
 ﻿//
-// Copyright 2019 Google LLC
+// Copyright 2023 Google LLC
 //
 // Licensed to the Apache Software Foundation (ASF) under one
 // or more contributor license agreements.  See the NOTICE file
@@ -21,141 +21,168 @@
 
 using Google.Apis.Auth.OAuth2;
 using Google.Solutions.Apis.Auth;
+using Google.Solutions.Apis.Auth.Gaia;
 using Google.Solutions.Common.Util;
+using Google.Solutions.IapDesktop.Application.Properties;
+using Google.Solutions.Platform.Net;
 using System;
+using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace Google.Solutions.IapDesktop.Application.Profile.Auth
 {
-    public class Authorization : IAuthorization
+    internal class Authorization : IAuthorization // TODO: Rename class
     {
-        private readonly ISignInClient adapter;
+        private IOidcClient client;
+        private IOidcSession session = null;
 
-        private readonly UserCredential credential;
-
-        private Authorization(
-            ISignInClient adapter,
-            IDeviceEnrollment deviceEnrollment,
-            UserCredential credential,
-            UserInfo userInfo)
+        private void SetOrSpliceSession(IOidcSession newSession)
         {
-            //
-            // The Authorization object is initialized during startup and
-            // a single instance is kept throughout the lifetime of the
-            // process, even in the case of re-auth.
-            //
-            // We must use the same UserCredential object throghout the
-            // lifetime of the app because existing clients maintain a 
-            // reference to the object.
-            //
-            // In case of re-auth, we therefore don't swap the credential
-            // object, but merely replace its embedded refresh token.
-            //
-            this.adapter = adapter.ExpectNotNull(nameof(adapter));
-            this.DeviceEnrollment = deviceEnrollment.ExpectNotNull(nameof(deviceEnrollment));
-            this.credential = credential.ExpectNotNull(nameof(credential));
-            this.UserInfo = userInfo.ExpectNotNull(nameof(userInfo));
-        }
+            newSession.ExpectNotNull(nameof(newSession));   
 
-        public event EventHandler Reauthorized;
-
-        public ICredential Credential => this.credential;
-
-        public string Email => this.UserInfo.Email;
-
-        public UserInfo UserInfo { get; private set; }
-
-        public IDeviceEnrollment DeviceEnrollment { get; }
-
-        public Task RevokeAsync()
-        {
-            return this.adapter.DeleteRefreshTokenAsync();
-        }
-
-        public async Task ReauthorizeAsync(CancellationToken token)
-        {
-            //
-            // As this is a 3p OAuth app, we do not support Gnubby/Password-based
-            // reauth. Instead, we simply trigger a new authorization (code flow).
-            //
-            // Use the current user as login hint to simplify the browser flow
-            // a little.
-            //
-            var newCredential = await this.adapter
-                .SignInWithBrowserAsync(this.Email, token)
-                .ConfigureAwait(false);
-
-            //
-            // Keep the credential object, but swap out its refresh token.
-            //
-            this.credential.Token = newCredential.Token;
-
-            //
-            // The user might have changed to a different user account,
-            // so we have to re-fetch user information.
-            //
-            this.UserInfo = await this.adapter
-                .QueryUserInfoAsync(newCredential, token)
-                .ConfigureAwait(false);
-
-            this.Reauthorized?.Invoke(this, EventArgs.Empty);
-        }
-
-        //---------------------------------------------------------------------
-        // Factory methods.
-        //---------------------------------------------------------------------
-
-        public static async Task<Authorization> TryLoadExistingAuthorizationAsync(
-            ISignInClient oauthAdapter,
-            IDeviceEnrollment deviceEnrollment,
-            CancellationToken token)
-        {
-            var credential = await oauthAdapter
-                .TrySignInWithRefreshTokenAsync(token)
-                .ConfigureAwait(false);
-            if (credential != null)
+            if (this.session != null)
             {
                 //
-                // Authorize worked, so the token was still valid.
+                // Once we have a session, we must never replace it,
+                // all we can do it is splice it to extend its lifetime.
                 //
-                var userInfo = await oauthAdapter
-                    .QueryUserInfoAsync(credential, token)
-                    .ConfigureAwait(false);
+                this.session.Splice(newSession);
 
-                return new Authorization(
-                    oauthAdapter,
-                    deviceEnrollment,
-                    credential,
-                    userInfo);
+                this.Reauthorized?.Invoke(this, EventArgs.Empty);
             }
             else
             {
-                //
-                // No token found, or it was invalid.
-                //
-                return null;
+                this.session = newSession;
             }
         }
 
-        public static async Task<Authorization> CreateAuthorizationAsync(
-            ISignInClient oauthAdapter,
-            IDeviceEnrollment deviceEnrollment,
-            CancellationToken token)
+        public Authorization(IOidcClient client)
         {
-            var credential = await oauthAdapter
-                .SignInWithBrowserAsync(null, token)
+            this.client = client.ExpectNotNull(nameof(client));
+        }
+
+        //---------------------------------------------------------------------
+        // IAuthorization.
+        //---------------------------------------------------------------------
+
+        public event EventHandler Reauthorized;
+
+        public IOidcSession Session
+        {
+            get => this.session ?? throw new InvalidOperationException("Not authorized yet");
+        }
+
+        //---------------------------------------------------------------------
+        // Publics.
+        //---------------------------------------------------------------------
+
+        /// <summary>
+        /// Raised before reauthorizations. Subscribers can use this event
+        /// to show a UI and/or modify the browser preference.
+        /// </summary>
+        public event EventHandler<ReauthorizeEventArgs> BeforeReauthorize;
+
+        /// <summary>
+        /// Try to authorize using an existing refresh token.
+        /// </summary>
+        public async Task<bool> TryAuthorizeSilentlyAsync(
+            CancellationToken cancellationToken)
+        {
+            Debug.Assert(
+                this.session == null,
+                "Silent authorize should only be performed initially");
+
+            var newSession = await this.client
+                .TryAuthorizeSilentlyAsync(cancellationToken)
                 .ConfigureAwait(false);
 
-            var userInfo = await oauthAdapter
-                .QueryUserInfoAsync(credential, token)
-                .ConfigureAwait(false);
+            if (newSession != null)
+            {
+                SetOrSpliceSession(newSession);
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+        }
 
-            return new Authorization(
-                oauthAdapter,
-                deviceEnrollment,
-                credential,
-                userInfo);
+        /// <summary>
+        /// Authorize or re-authorize using a browser-based OIDC flow.
+        /// </summary>
+        public async Task AuthorizeAsync(
+            BrowserPreference browserPreference,
+            CancellationToken cancellationToken)
+        {
+            var newSession = await this.client
+                .AuthorizeAsync(
+                    new BrowserCodeReceiver(browserPreference),
+                    cancellationToken)
+                .ConfigureAwait(false);
+            SetOrSpliceSession(newSession);
+        }
+
+        public Task ReauthorizeAsync(CancellationToken cancellationToken)
+        {
+            Debug.Assert(this.session != null);
+
+            //
+            // Give the UI a chance to react and decide which browser to use.
+            //
+            var args = new ReauthorizeEventArgs();
+            this.BeforeReauthorize?.Invoke(this, args);
+
+            return AuthorizeAsync(args.BrowserPreference, cancellationToken);
+        }
+
+        //---------------------------------------------------------------------
+        // Inner classes.
+        //---------------------------------------------------------------------
+
+        public class ReauthorizeEventArgs : EventArgs
+        {
+            public BrowserPreference BrowserPreference { get; set; } 
+                = BrowserPreference.Default;
+        }
+
+        private class BrowserCodeReceiver : LocalServerCodeReceiver 
+        {
+            private readonly BrowserPreference browserPreference;
+
+            public BrowserCodeReceiver(BrowserPreference browserPreference)
+                : base(Resources.AuthorizationSuccessful)
+            {
+                this.browserPreference = browserPreference;
+            }
+
+            protected override bool OpenBrowser(string url)
+            {
+                Browser
+                    .Get(this.browserPreference)
+                    .Navigate(url);
+                return true;
+            }
+        }
+
+        //---------------------------------------------------------------------
+        // TODO: Obsolete methods below?
+
+        public ICredential Credential => this.Session.ApiCredential;
+
+        public string Email => this.Session.Username;
+
+        public UserInfo UserInfo => new UserInfo()
+        {
+            Email = this.Session.Username,
+            HostedDomain = (this.Session as IGaiaOidcSession)?.HostedDomain
+        };
+
+        public IDeviceEnrollment DeviceEnrollment => this.Session.DeviceEnrollment;
+
+        public Task RevokeAsync()
+        {
+            throw new NotImplementedException();
         }
     }
 }
